feat: 🎸 add chat pages (#44)
* feat: 🎸 add chat pages

* feat: 🎸 remove fake time
