+++ src/App.jsx
@@ -9,7 +9,10 @@ import { motion } from 'framer-motion';
 import { BgBubble } from './components';
 import Home from './views';
 import Role from './views/introduction/Role';
-import ChatSprintDailyScrum from './views/chat/SprintDailyScrum';
+// import ChatSprintDailyScrum from './views/chat/SprintDailyScrum';
+import {
+  Chat, SprintDailyScrum, SprintReview, SprintRetro,
+} from './views/chat';
 import './App.css';
 import ExamSprintPoint from './views/exam/SprintPoint';
 import { Scrum, IntroProductBacklog } from './views/introduction';
@@ -58,10 +61,11 @@ function App() {
           <Route element={<AnimationLayout />}>
             <Route path="/" element={<Home />} />
             <Route path="/exam/sprint-point" element={<ExamSprintPoint />} />
-            <Route
-              path="/chat/sprint-daily-scrum"
-              element={<ChatSprintDailyScrum />}
-            />
+            <Route path="/chat" element={<Chat />}>
+              <Route path="sprint-daily-scrum" element={<SprintDailyScrum />} />
+              <Route path="sprint-review" element={<SprintReview />} />
+              <Route path="sprint-retro" element={<SprintRetro />} />
+            </Route>
             <Route path="/introduction/scrum" element={<Scrum />} />
             <Route path="/introduction/role" element={<Role />} />
             <Route path="/introduction/product-backlog" element={<IntroProductBacklog />} />
