+++ src/components/chat/ChatNavBar.jsx
@@ -1,13 +1,8 @@
 import clsx from 'clsx';
-import { useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { Icons } from '../Icons';
 
 export function ChatNavBar({ previousPage }) {
-  const navigate = useNavigate();
-  const goBack = () => {
-    const preRoute = '/';
-    navigate(preRoute);
-  };
   return (
     <div className={clsx(
       'flex flex-row items-center gap-3',
@@ -15,14 +10,15 @@ export function ChatNavBar({ previousPage }) {
       'bg-teal-500',
     )}
     >
-      <button
-        type="button"
-        className="w-2"
-        onClick={goBack}
-      >
-        <Icons.ArrowLeft className="text-white" />
-      </button>
-      <p className="inline-block text-white text-xs">{ previousPage }</p>
+      <Link to={previousPage}>
+        <button
+          type="button"
+          className="w-2"
+        >
+          <Icons.ArrowLeft className="text-white" />
+        </button>
+      </Link>
+      <p className="inline-block text-white text-xs">Scrum 新手村專案</p>
     </div>
   );
 }
