+++ src/components/index.js
@@ -4,3 +4,4 @@ export * from './Button';
 export * from './BlurBlockBg';
 export * from './chat';
 export * from './introduction';
+export * from './ChatRoom';
