+++ src/views/chat/SprintRetro.jsx
@@ -1,0 +1,10 @@
+import { chatRetroInfo } from '../../constants/chatRetroInfo';
+import { ChatRoom } from '../../components';
+
+export function SprintRetro() {
+  return (
+    <ChatRoom props={chatRetroInfo} />
+  );
+}
+
+export default SprintRetro;
