+++ src/views/chat/SprintReview.jsx
@@ -1,0 +1,10 @@
+import { chatReviewInfo } from '../../constants/chatReviewInfo';
+import { ChatRoom } from '../../components';
+
+export function SprintReview() {
+  return (
+    <ChatRoom props={chatReviewInfo} />
+  );
+}
+
+export default SprintReview;
